+++ Backgrounds/InfernoSky.cs
@@ -73,7 +73,7 @@ namespace AAMod.Backgrounds
                     {
                         num64 = 0f;
                     }
-                    int num20 = (int)(Main.time / 54000.0 * (double)(Main.screenWidth + Main.sunTexture.Width * 2)) - Main.sunTexture.Width;
+                    int num20 = (int)(Main.time / 54000.0 * (Main.screenWidth + Main.sunTexture.Width * 2)) - Main.sunTexture.Width;
                     int num21 = 0;
                     float num22 = 1f;
                     float rotation = (float)(Main.time / 54000.0) * 2f - 7.3f;
@@ -128,9 +128,9 @@ namespace AAMod.Backgrounds
                     return;
                 }
                 float scale = Math.Min(1f, (Main.screenPosition.Y - 1000f) / 1000f);
-                Vector2 value3 = Main.screenPosition + new Vector2((float)(Main.screenWidth >> 1), (float)(Main.screenHeight >> 1));
+                Vector2 value3 = Main.screenPosition + new Vector2(Main.screenWidth >> 1, Main.screenHeight >> 1);
                 Rectangle rectangle = new Rectangle(-1000, -1000, 4000, 4000);
-                for (int j = (int)num; j < num2; j++)
+                for (int j = num; j < num2; j++)
                 {
                     Vector2 value4 = new Vector2(1f / Meteors[j].Depth, 0.9f / Meteors[j].Depth);
                     Vector2 position = (Meteors[j].Position - value3) * value4 + value3 - Main.screenPosition;
@@ -156,7 +156,7 @@ namespace AAMod.Backgrounds
             Meteors = new Meteor[150];
             for (int i = 0; i < Meteors.Length; i++)
             {
-                float num = (float)i / (float)Meteors.Length;
+                float num = i / (float)Meteors.Length;
                 Meteors[i].Position.X = num * (Main.maxTilesX * 16f) + this._random.NextFloat() * 40f - 20f;
                 Meteors[i].Position.Y = this._random.NextFloat() * -((float)Main.worldSurface * 16f + 10000f) - 10000f;
                 if (this._random.Next(3) != 0)
