+++ Backgrounds/ScreenClouds.cs
@@ -76,11 +76,11 @@ namespace AAMod.Backgrounds
 		public Color GetAlpha(Color newColor, float alph)
 		{
 			int alpha = 255 - (int)(255 * alph);
-			float alphaDiff = (float)(255 - alpha) / 255f;
-			int newR = (int)((float)newColor.R * alphaDiff);
-			int newG = (int)((float)newColor.G * alphaDiff);
-			int newB = (int)((float)newColor.B * alphaDiff);
-			int newA = (int)newColor.A - alpha;
+			float alphaDiff = (255 - alpha) / 255f;
+			int newR = (int)(newColor.R * alphaDiff);
+			int newG = (int)(newColor.G * alphaDiff);
+			int newB = (int)(newColor.B * alphaDiff);
+			int newA = newColor.A - alpha;
 			if (newA < 0) newA = 0;
 			if (newA > 255) newA = 255;		
 			return new Color(newR, newG, newB, newA);
