+++ Buffs/Sharkron.cs
@@ -21,7 +21,7 @@ namespace AAMod.Buffs
 			bool petProjectileNotSpawned = player.ownedProjectileCounts[mod.ProjectileType("Sharkron")] <= 0;
 			if (petProjectileNotSpawned && player.whoAmI == Main.myPlayer)
 			{
-				Projectile.NewProjectile(player.position.X + (float)(player.width / 2), player.position.Y + (float)(player.height / 2), 0f, 0f, mod.ProjectileType("Sharkron"), 0, 0f, player.whoAmI, 0f, 0f);
+				Projectile.NewProjectile(player.position.X + player.width / 2, player.position.Y + player.height / 2, 0f, 0f, mod.ProjectileType("Sharkron"), 0, 0f, player.whoAmI, 0f, 0f);
 			}
 		}
 	}
