+++ Globals/AAAI.cs
@@ -19,26 +19,26 @@ namespace AAMod
             {
                 xVelocityChanged = true;
             }
-            if (npc.position.X == npc.oldPosition.X || ai[3] >= (float)ticksUntilBoredom || xVelocityChanged)
+            if (npc.position.X == npc.oldPosition.X || ai[3] >= ticksUntilBoredom || xVelocityChanged)
             {
                 ai[3] += 1f;
             }
             else
-            if ((double)Math.Abs(npc.velocity.X) > 0.9 && ai[3] > 0f) { ai[3] -= 1f; }
-            if (ai[3] > (float)(ticksUntilBoredom * 10)) { ai[3] = 0f; }
+            if (Math.Abs(npc.velocity.X) > 0.9 && ai[3] > 0f) { ai[3] -= 1f; }
+            if (ai[3] > ticksUntilBoredom * 10) { ai[3] = 0f; }
             if (npc.justHit) { ai[3] = 0f; }
-            if (ai[3] == (float)ticksUntilBoredom) { npc.netUpdate = true; }
+            if (ai[3] == ticksUntilBoredom) { npc.netUpdate = true; }
 
-            bool notBored = ai[3] < (float)ticksUntilBoredom;
+            bool notBored = ai[3] < ticksUntilBoredom;
             //if npc does not flee when it's day, if is night, or npc is not on the surface and it hasn't updated projectile pass, update target.
-            if (targetPlayers && (!fleeWhenNight || Main.dayTime || (double)npc.position.Y > Main.worldSurface * 16.0) && (fleeWhenNight && Main.dayTime ? notBored : (!allowBoredom || notBored)))
+            if (targetPlayers && (!fleeWhenNight || Main.dayTime || npc.position.Y > Main.worldSurface * 16.0) && (fleeWhenNight && Main.dayTime ? notBored : (!allowBoredom || notBored)))
             {
                 npc.TargetClosest(true);
             }
             else
             if (ai[2] <= 0f)//if 'bored'
             {
-                if (fleeWhenNight && !Main.dayTime && (double)(npc.position.Y / 16f) < Main.worldSurface && npc.timeLeft > 10)
+                if (fleeWhenNight && !Main.dayTime && npc.position.Y / 16f < Main.worldSurface && npc.timeLeft > 10)
                 {
                     npc.timeLeft = 10;
                 }
@@ -268,7 +268,7 @@ namespace AAMod
                     float timerPartial = 1f - npc.localAI[0] / timerMax;
                     float timerPartialTimes20 = timerPartial * 20f;
                     int nextNPC = 0;
-                    while ((float)nextNPC < timerPartialTimes20)
+                    while (nextNPC < timerPartialTimes20)
                     {
                         npcAvoidCollision = nextNPC;
                         nextNPC = npcAvoidCollision + 1;
@@ -279,7 +279,7 @@ namespace AAMod
             {
                 npc.TargetClosest(true);
                 npc.ai[0] = 1f;
-                npc.ai[1] = (float)npc.direction;
+                npc.ai[1] = npc.direction;
             }
             else if (npc.ai[0] == 1f)
             {
@@ -322,7 +322,7 @@ namespace AAMod
                 {
                     npc.TargetClosest(true);
                     npc.ai[0] = 3f;
-                    npc.ai[1] = (float)npc.direction;
+                    npc.ai[1] = npc.direction;
                 }
             }
             else if (npc.ai[0] == 3f)
@@ -341,7 +341,7 @@ namespace AAMod
                 {
                     npc.TargetClosest(true);
                     npc.ai[0] = 0f;
-                    npc.ai[1] = (float)npc.direction;
+                    npc.ai[1] = npc.direction;
                 }
             }
         }
@@ -443,26 +443,26 @@ namespace AAMod
             {
                 xVelocityChanged = true;
             }
-            if (npc.position.X == npc.oldPosition.X || ai[3] >= (float)ticksUntilBoredom || xVelocityChanged)
+            if (npc.position.X == npc.oldPosition.X || ai[3] >= ticksUntilBoredom || xVelocityChanged)
             {
                 ai[3] += 1f;
             }
             else
-            if ((double)Math.Abs(npc.velocity.X) > 0.9 && ai[3] > 0f) { ai[3] -= 1f; }
-            if (ai[3] > (float)(ticksUntilBoredom * 10)) { ai[3] = 0f; }
+            if (Math.Abs(npc.velocity.X) > 0.9 && ai[3] > 0f) { ai[3] -= 1f; }
+            if (ai[3] > ticksUntilBoredom * 10) { ai[3] = 0f; }
             if (npc.justHit) { ai[3] = 0f; }
-            if (ai[3] == (float)ticksUntilBoredom) { npc.netUpdate = true; }
+            if (ai[3] == ticksUntilBoredom) { npc.netUpdate = true; }
 
-            bool notBored = ai[3] < (float)ticksUntilBoredom;
+            bool notBored = ai[3] < ticksUntilBoredom;
             //if npc does not flee when it's day, if is night, or npc is not on the surface and it hasn't updated projectile pass, update target.
-            if (targetPlayers && (player.ZoneCorrupt || (double)npc.position.Y > Main.worldSurface * 16.0) && (Main.dayTime ? notBored : (!allowBoredom || notBored)))
+            if (targetPlayers && (player.ZoneCorrupt || npc.position.Y > Main.worldSurface * 16.0) && (Main.dayTime ? notBored : (!allowBoredom || notBored)))
             {
                 npc.TargetClosest(true);
             }
             else
             if (ai[2] <= 0f)//if 'bored'
             {
-                if (!player.ZoneCorrupt && (double)(npc.position.Y / 16f) < Main.worldSurface && npc.timeLeft > 10)
+                if (!player.ZoneCorrupt && npc.position.Y / 16f < Main.worldSurface && npc.timeLeft > 10)
                 {
                     npc.timeLeft = 10;
                 }
@@ -533,7 +533,7 @@ namespace AAMod
         public static void GripAI(NPC npc, ref float[] ai, float moveIntervalX = 0.1f, float moveIntervalY = 0.04f, float velMaxX = 4f, float velMaxY = 1.5f)
         {
             //if it should flee when it's day, and it is day, the npc's position is at or above the surface, it will flee.
-            if (Main.dayTime && (double)npc.position.Y <= Main.worldSurface * 16.0)
+            if (Main.dayTime && npc.position.Y <= Main.worldSurface * 16.0)
             {
                 if (npc.timeLeft > 10) { npc.timeLeft = 10; }
                 npc.directionY = -1;
