+++ Globals/AAColor.cs
@@ -43,7 +43,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(224, 213, 94), new Color(130, 104, 41), new Color(224, 213, 94));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(224, 213, 94), new Color(130, 104, 41), new Color(224, 213, 94));
             }
         }
         
@@ -51,7 +51,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.LightCoral, Color.Coral, Color.LightCoral);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.LightCoral, Color.Coral, Color.LightCoral);
             }
         }
 
@@ -59,7 +59,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.Orange, Color.DarkOrange, Color.Orange);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.Orange, Color.DarkOrange, Color.Orange);
             }
         }
 
@@ -139,7 +139,7 @@ namespace AAMod
 		{
 			get
 			{
-				return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.Red, Color.DarkRed, Color.Red);
+				return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.Red, Color.DarkRed, Color.Red);
 			}
 		}
 
@@ -155,7 +155,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.OrangeRed, Color.Indigo, Color.Indigo, Color.OrangeRed);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.OrangeRed, Color.Indigo, Color.Indigo, Color.OrangeRed);
             }
         }
 
@@ -163,7 +163,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Shen, AkumaA, Shen, YamataA, Shen, AkumaA, Shen);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Shen, AkumaA, Shen, YamataA, Shen, AkumaA, Shen);
             }
         }
 
@@ -171,7 +171,7 @@ namespace AAMod
         { 
 			get
 			{
-				return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.DarkGray, Color.Gray, Color.DarkGray);
+				return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.DarkGray, Color.Gray, Color.DarkGray);
 			}
 		}
 
@@ -179,7 +179,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.Cyan, Color.DarkCyan, Color.Cyan);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.Cyan, Color.DarkCyan, Color.Cyan);
             }
         }
 
@@ -187,7 +187,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(10, 22, 28), Color.White, new Color(10, 22, 28));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(10, 22, 28), Color.White, new Color(10, 22, 28));
             }
         }
 
@@ -206,7 +206,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(0, 140, 128), new Color(0, 117, 128), new Color(0, 140, 128));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(0, 140, 128), new Color(0, 117, 128), new Color(0, 140, 128));
             }
         }
 
@@ -214,7 +214,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(227, 159, 0), new Color(176, 88, 0), new Color(227, 159, 0));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(227, 159, 0), new Color(176, 88, 0), new Color(227, 159, 0));
             }
         }
 
@@ -222,7 +222,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(112, 255, 243), new Color(27, 178, 191), new Color(112, 255, 243));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(112, 255, 243), new Color(27, 178, 191), new Color(112, 255, 243));
             }
         }
 
@@ -230,7 +230,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(179, 75, 0), new Color(128, 43, 0), new Color(179, 75, 0));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(179, 75, 0), new Color(128, 43, 0), new Color(179, 75, 0));
             }
         }
 
@@ -238,7 +238,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(214, 36, 0), new Color(153, 13, 0), new Color(214, 36, 0));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(214, 36, 0), new Color(153, 13, 0), new Color(214, 36, 0));
             }
         }
 
@@ -246,7 +246,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(57, 0, 156), new Color(26, 0, 128), new Color(57, 0, 156));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(57, 0, 156), new Color(26, 0, 128), new Color(57, 0, 156));
             }
         }
 
@@ -254,7 +254,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(232, 36, 102), new Color(33, 31, 77), new Color(232, 36, 102));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(232, 36, 102), new Color(33, 31, 77), new Color(232, 36, 102));
             }
         }
 
@@ -262,7 +262,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(117, 0, 140), new Color(75, 0, 102), new Color(117, 0, 140));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(117, 0, 140), new Color(75, 0, 102), new Color(117, 0, 140));
             }
         }
 
@@ -270,7 +270,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(140, 0, 2), new Color(102, 0, 2), new Color(140, 0, 2));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(140, 0, 2), new Color(102, 0, 2), new Color(140, 0, 2));
             }
         }
 
@@ -278,7 +278,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(124, 140, 2), new Color(56, 102, 0), new Color(124, 140, 2));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(124, 140, 2), new Color(56, 102, 0), new Color(124, 140, 2));
             }
         }
 
@@ -286,7 +286,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(0, 138, 166), new Color(0, 99, 148), new Color(0, 138, 166));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(0, 138, 166), new Color(0, 99, 148), new Color(0, 138, 166));
             }
         }
 
@@ -294,7 +294,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(122, 47, 24), new Color(77, 22, 15), new Color(122, 47, 24));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(122, 47, 24), new Color(77, 22, 15), new Color(122, 47, 24));
             }
         }
 
@@ -302,7 +302,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.DarkCyan, Color.Magenta, Color.DarkCyan);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.DarkCyan, Color.Magenta, Color.DarkCyan);
             }
         }
 
@@ -310,7 +310,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.DarkGreen, Color.Lime, Color.DarkGreen);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.DarkGreen, Color.Lime, Color.DarkGreen);
             }
         }
 
@@ -336,11 +336,11 @@ namespace AAMod
             {
                 if (Main.dayTime && Celestial != Daybringer)
                 {
-                    return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Nightcrawler, Daybringer);
+                    return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Nightcrawler, Daybringer);
                 }
                 else if (!Main.dayTime && Celestial != Nightcrawler)
                 {
-                    return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Daybringer, Nightcrawler);
+                    return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Daybringer, Nightcrawler);
                 }
                 if (Main.dayTime)
                 {
@@ -357,7 +357,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.DarkViolet, Color.DarkViolet, Color.DarkViolet, Color.DarkViolet, Color.DarkViolet, Color.Violet, Color.DarkViolet, Color.Violet, Color.DarkViolet);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.DarkViolet, Color.DarkViolet, Color.DarkViolet, Color.DarkViolet, Color.DarkViolet, Color.Violet, Color.DarkViolet, Color.Violet, Color.DarkViolet);
             }
         }
         
@@ -376,7 +376,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.SandyBrown, Color.Sienna, Color.SandyBrown);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.SandyBrown, Color.Sienna, Color.SandyBrown);
             }
         }
         
@@ -384,7 +384,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.CornflowerBlue, Color.CadetBlue, Color.CornflowerBlue);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.CornflowerBlue, Color.CadetBlue, Color.CornflowerBlue);
             }
         }
         
@@ -392,7 +392,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.DodgerBlue, Color.CornflowerBlue, Color.DodgerBlue);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.DodgerBlue, Color.CornflowerBlue, Color.DodgerBlue);
             }
         }
 
@@ -400,7 +400,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.Violet, Color.ForestGreen, Color.ForestGreen, Color.Violet);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.Violet, Color.ForestGreen, Color.ForestGreen, Color.Violet);
             }
         }
 
@@ -408,7 +408,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.SeaGreen, Color.DarkSeaGreen, Color.SeaGreen);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.SeaGreen, Color.DarkSeaGreen, Color.SeaGreen);
             }
         }
 
@@ -416,7 +416,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.Red, Color.DarkRed, Color.DarkRed, Color.Red);
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.Red, Color.DarkRed, Color.DarkRed, Color.Red);
             }
         }
 
@@ -424,7 +424,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(55, 200, 26), new Color(179, 252, 0), new Color(218, 255, 0), new Color(179, 252, 0), new Color(55, 200, 26));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(55, 200, 26), new Color(179, 252, 0), new Color(218, 255, 0), new Color(179, 252, 0), new Color(55, 200, 26));
             }
         }
 
@@ -432,7 +432,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(253, 152, 0), new Color(252, 202, 80), new Color(255, 251, 166), new Color(252, 202, 80), new Color(253, 152, 0));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(253, 152, 0), new Color(252, 202, 80), new Color(255, 251, 166), new Color(252, 202, 80), new Color(253, 152, 0));
             }
         }
 
@@ -440,7 +440,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(32, 51, 91), new Color(42, 120, 165), new Color(32, 51, 91));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(32, 51, 91), new Color(42, 120, 165), new Color(32, 51, 91));
             }
         }
 
@@ -448,7 +448,7 @@ namespace AAMod
         {
             get
             {
-                return BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, new Color(125, 10, 50), new Color(212, 45, 40), new Color(125, 10, 50));
+                return BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, new Color(125, 10, 50), new Color(212, 45, 40), new Color(125, 10, 50));
             }
         }
 
@@ -456,7 +456,7 @@ namespace AAMod
 		{
 			get
 			{
-				Color c = BaseUtility.MultiLerpColor((float)(Main.player[Main.myPlayer].miscCounter % 100) / 100f, Color.White, Color.White * 0.7f, Color.White);
+				Color c = BaseUtility.MultiLerpColor(Main.player[Main.myPlayer].miscCounter % 100 / 100f, Color.White, Color.White * 0.7f, Color.White);
 				c.A = 255;
 				return c;
 			}
